+++ services/command-router-quarkus/src/main/java/org/eclipse/hono/commandrouter/quarkus/Application.java
@@ -180,12 +180,14 @@ public class Application extends AbstractServiceApplication {
         this.kafkaProducerConfig = new MessagingKafkaProducerConfigProperties();
         this.kafkaProducerConfig.setCommonClientConfig(options.commonClientConfig());
         this.kafkaProducerConfig.setProducerConfig(options.producerConfig());
-        this.kafkaProducerConfig.setDefaultClientIdPrefix(DEFAULT_CLIENT_ID_PREFIX);
+        this.kafkaProducerConfig.setDefaultClientIdPrefix(
+                options.defaultClientIdPrefix().orElse(DEFAULT_CLIENT_ID_PREFIX));
 
         this.kafkaConsumerConfig = new MessagingKafkaConsumerConfigProperties();
         this.kafkaConsumerConfig.setCommonClientConfig(options.commonClientConfig());
         this.kafkaConsumerConfig.setConsumerConfig(options.consumerConfig());
-        this.kafkaConsumerConfig.setDefaultClientIdPrefix(DEFAULT_CLIENT_ID_PREFIX);
+        this.kafkaConsumerConfig.setDefaultClientIdPrefix(
+                options.defaultClientIdPrefix().orElse(DEFAULT_CLIENT_ID_PREFIX));
     }
 
     @Override